Transaction bug solved

+++ app/src/main/java/com/engineersapps/eapps/models/payment/CoursePaymentRequest.kt
@@ -2,4 +2,4 @@ package com.engineersapps.eapps.models.payment
 
 data class CoursePaymentRequest(val mobile: String?,var invoiceid: String = "", val student_id: Int?,
                                 val course_id: Int?, val total_amount: Int?,
-                                val paid_amount: Int?, val duration: Int?)
+                                val paid_amount: Double?, val duration: Int?)
